requirements.txt
line-bot-sdk
numpy
yahoo_fin
bs4
flask
time
pymongo
datetime
pandas
SnowNLP
emoji
pyshorteners
scipy
schedule
